\name{similarity}
\alias{similarity}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ calculate the correlation of two vector}
\description{
  calculate the correlation of two vector
}
\usage{
similarity(target,x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{target}{vector 1}
  \item{x}{vector2}
}

\value{
The return is similarity
}


\author{Yizeng Liang ,Zhang Zhimin, Chen Shan}

\examples{
	data(simulate)
	similarity(p1,p2)
}
\keyword{ methods }
